refactor(annomate): removed legacy annomate and microsentry
